Simplified collectors
Collectors are basically response handlers that are queued at the same
time as outbound message in a separate queue. Later, when response
arrives next collector is popped from the queue to handle it.
Collector interface was quite wide and allowed processing parts of the
SUCCESS message metadata.

This commit turns `Collector` into `ResponseHandler` and makes it only
expose three methods to handle SUCCESS, FAILURE and RECORD messages.